refactor(desktop): move preset harness data into discovery/presets.rs
discovery.rs is 1860 lines — over the 1000-line desktop file-size limit,
so the ratchet in `pnpm check` forbids any growth, and adding a preset
entry fails it. Move the tier-2 data (PresetHarness, PRESET_HARNESSES,
and the two helpers projecting it) into a sibling module, which drops
discovery.rs to ~1716 lines and gives presets a small file of their own
to grow in.

Pure code motion: visibility widens to pub(super) for the fields
discovery.rs reads, and the presetLogos guard now parses presets.rs.
